mrf_final_output/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyspark>=3.5.0
pyyaml>=6.0
python-multipart>=0.0.6
